Split sensor code out into their own files
