+++ include/BLE_Common.h
@@ -201,181 +201,3 @@ private:
 };
 
 extern SpinBLEClient spinBLEClient;
-
-class SensorData
-{
-public:
-    SensorData(String id) : id(id){};
-
-    String getId();
-    virtual bool hasHeartRate() = 0;
-    virtual bool hasCadence() = 0;
-    virtual bool hasPower() = 0;
-    virtual bool hasSpeed() = 0;
-    virtual int getHeartRate() = 0;
-    virtual float getCadence() = 0;
-    virtual int getPower() = 0;
-    virtual float getSpeed() = 0;
-    virtual void decode(uint8_t *data, size_t length) = 0;
-
-protected:
-    String id;
-};
-
-class KnownDevice
-{
-public:
-    KnownDevice(NimBLEUUID uuid, std::shared_ptr<SensorData> sensorData): uuid(uuid), sensorData(sensorData){};
-    NimBLEUUID getUUID();
-    std::shared_ptr<SensorData> decode(uint8_t *data, size_t length);
-    
-private:
-    NimBLEUUID uuid;
-    std::shared_ptr<SensorData> sensorData;
-};
-
-class SensorDataFactory
-{
-public:
-    SensorDataFactory(){};
-        
-    std::shared_ptr<SensorData> getSensorData(BLERemoteCharacteristic *characteristic, uint8_t *data, size_t length);
-
-private:
-    std::vector<KnownDevice*> knownDevices;
-    static std::shared_ptr<SensorData> NULL_SENSOR_DATA;
-};
-
-class NullData : public SensorData
-{
-public:
-    NullData() : SensorData("Null"){};
-
-    virtual bool hasHeartRate();
-    virtual bool hasCadence();
-    virtual bool hasPower();
-    virtual bool hasSpeed();
-    virtual int getHeartRate();
-    virtual float getCadence();
-    virtual int getPower();
-    virtual float getSpeed();
-    virtual void decode(uint8_t *data, size_t length);
-};
-
-class HeartRateData : public SensorData
-{
-public:
-    HeartRateData() : SensorData("HRM"){};
-
-    virtual bool hasHeartRate();
-    virtual bool hasCadence();
-    virtual bool hasPower();
-    virtual bool hasSpeed();
-    virtual int getHeartRate();
-    virtual float getCadence();
-    virtual int getPower();
-    virtual float getSpeed();
-    virtual void decode(uint8_t *data, size_t length);
-
-private:
-    int heartrate;
-};
-
-class CyclePowerData : public SensorData
-{
-public:
-    CyclePowerData() : SensorData("CPS"){};
-
-    virtual bool hasHeartRate();
-    virtual bool hasCadence();
-    virtual bool hasPower();
-    virtual bool hasSpeed();
-    virtual int getHeartRate();
-    virtual float getCadence();
-    virtual int getPower();
-    virtual float getSpeed();
-    virtual void decode(uint8_t *data, size_t length);
-
-private:
-    int power;
-    float cadence;
-    float crankRev = 0;
-    float lastCrankRev = 0;
-    float lastCrankEventTime = 0;
-    float crankEventTime = 0;
-    uint8_t missedReadingCount = 0;
-};
-
-class FlywheelData : public SensorData
-{
-public:
-    FlywheelData() : SensorData("FLYW"){};
-
-    virtual bool hasHeartRate();
-    virtual bool hasCadence();
-    virtual bool hasPower();
-    virtual bool hasSpeed();
-    virtual int getHeartRate();
-    virtual float getCadence();
-    virtual int getPower();
-    virtual float getSpeed();
-    virtual void decode(uint8_t *data, size_t length);
-
-private:
-    bool hasData = false;
-    float cadence;
-    int power;
-};
-
-class FitnessMachineIndoorBikeData : public SensorData
-{
-public:
-    FitnessMachineIndoorBikeData() : SensorData("FTMS"){};
-
-    enum Types : uint8_t
-    {
-        InstantaneousSpeed = 0,
-        AverageSpeed = 1,
-        InstantaneousCadence = 2,
-        AverageCadence = 3,
-        TotalDistance = 4,
-        ResistanceLevel = 5,
-        InstantaneousPower = 6,
-        AveragePower = 7,
-        TotalEnergy = 8,
-        EnergyPerHour = 9,
-        EnergyPerMinute = 10,
-        HeartRate = 11,
-        MetabolicEquivalent = 12,
-        ElapsedTime = 13,
-        RemainingTime = 14
-    };
-
-    static constexpr uint8_t FieldCount = Types::RemainingTime + 1;
-
-    virtual bool hasHeartRate();
-    virtual bool hasCadence();
-    virtual bool hasPower();
-    virtual bool hasSpeed();
-    virtual int getHeartRate();
-    virtual float getCadence();
-    virtual int getPower();
-    virtual float getSpeed();
-    virtual void decode(uint8_t *data, size_t length);
-
-private:
-    double_t values[FieldCount];
-
-    static uint8_t const flagBitIndices[];
-    static uint8_t const flagEnabledValues[];
-    static size_t const byteSizes[];
-    static uint8_t const signedFlags[];
-    static double_t const resolutions[];
-    static int convert(int value, size_t length, uint8_t isSigned);
-};
-
-/******************CPS***************************/
-void BLE_CPSDecode(BLERemoteCharacteristic *pBLERemoteCharacteristic);
-
-/******************FTMS**************************/
-void BLE_FTMSDecode(BLERemoteCharacteristic *pBLERemoteCharacteristic);
