+++ src/sensors/SensorData.cpp
@@ -1,0 +1,3 @@
+#include "sensors/SensorData.h"
+
+String SensorData::getId() { return this->id; }
